feat: add route / to /staff
